Add support for supplemental groups
This commit adds the ability to define lists of users that can be
forced into certain groups on login.  This is useful in the case of an
installation where users are stored in an external authentication
source that synchronizes groups each time it is queried.  It is
strongly discouraged to use this feature if you have the ability to
create native groups.